Proposal Development - Modify the Activity Type dropdown list To actually Use an Active field so inactive values do not appear and also sort by a SortOrder Field so values in the dropdown can be ordered as desired.
Proposal Development - Modify the Proposal Type dropdown list To actually Use an Active field so inactive values do not appear and also sort by a SortOrder Field so values in the dropdown can be ordered as desired.

Proposal Development - Modify the Sponsor Deadline Type dropdown list To actually Use an Active field so inactive values do not appear and also sort by a SortOrder Field so values in the dropdown can be ordered as desired.